Pointless refactoring of models
