+++ api/src/main/kotlin/com/gmtkgamejam/models/bot/dtos/BotDmDto.kt
@@ -1,4 +1,4 @@
-package com.gmtkgamejam.models
+package com.gmtkgamejam.models.bot.dtos
 
 import kotlinx.serialization.Serializable
