+++ api/src/main/kotlin/com/gmtkgamejam/models/posts/Availability.kt
@@ -1,4 +1,4 @@
-package com.gmtkgamejam.models
+package com.gmtkgamejam.models.posts
 
 enum class Availability {
