+++ api/src/main/kotlin/com/gmtkgamejam/models/posts/FavouritesList.kt
@@ -1,4 +1,4 @@
-package com.gmtkgamejam.models
+package com.gmtkgamejam.models.posts
 
 import kotlinx.serialization.Serializable
