+++ api/src/main/kotlin/com/gmtkgamejam/services/AdminService.kt
@@ -1,6 +1,6 @@
 package com.gmtkgamejam.services
 
-import com.gmtkgamejam.models.BannedUser
+import com.gmtkgamejam.models.admin.BannedUser
 import com.gmtkgamejam.repositories.AdminRepository
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
