+++ api/src/main/kotlin/com/gmtkgamejam/services/AuthService.kt
@@ -1,6 +1,6 @@
 package com.gmtkgamejam.services
 
-import com.gmtkgamejam.models.AuthTokenSet
+import com.gmtkgamejam.models.auth.AuthTokenSet
 import com.mongodb.client.MongoClient
 import com.mongodb.client.MongoCollection
 import com.mongodb.client.model.UpdateOptions
