+++ api/src/main/kotlin/com/gmtkgamejam/services/FavouritesService.kt
@@ -1,7 +1,7 @@
 package com.gmtkgamejam.services
 
-import com.gmtkgamejam.models.FavouritePostDto
-import com.gmtkgamejam.models.FavouritesList
+import com.gmtkgamejam.models.posts.dtos.FavouritePostDto
+import com.gmtkgamejam.models.posts.FavouritesList
 import com.gmtkgamejam.repositories.FavouritesRepository
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
